# kvs: use json object to find duplicate keys