As discussed in #3615, there is some bottlenecking at larger scales in the KVS `normalize_and_append_unique()` function.  In the function, a json array is regularly searched to determine if any keys are duplicates of each other.  This is is an O(N^2) search with a not so friendly `strcmp()` in the middle.

As an experiment I modified the "array search" to use a json object as a "hash lookup".  Effectively replacing a `json_array_foreach` & `strcmp` with a single call to `json_object_get()`.

The results using the nsbench benchmark.

Before

```
  0.17s: Benchmark on 256 namespaces complete: 1468.3 namespace/s
  0.35s: Benchmark on 512 namespaces complete: 1469.0 namespace/s
  0.70s: Benchmark on 1024 namespaces complete: 1461.4 namespace/s
  1.35s: Benchmark on 2048 namespaces complete: 1519.2 namespace/s
  2.68s: Benchmark on 4096 namespaces complete: 1527.9 namespace/s
  5.31s: Benchmark on 8192 namespaces complete: 1543.5 namespace/s
 11.13s: Benchmark on 16384 namespaces complete: 1471.8 namespace/s
 22.50s: Benchmark on 32768 namespaces complete: 1456.6 namespace/s
 47.77s: Benchmark on 65536 namespaces complete: 1371.8 namespace/s
115.77s: Benchmark on 131072 namespaces complete: 1132.2 namespace/s
364.97s: Benchmark on 262144 namespaces complete: 718.3 namespace/s
```

After

```
  0.19s: Benchmark on 256 namespaces complete: 1379.0 namespace/s
  0.37s: Benchmark on 512 namespaces complete: 1401.2 namespace/s
  0.71s: Benchmark on 1024 namespaces complete: 1437.1 namespace/s
  1.34s: Benchmark on 2048 namespaces complete: 1526.3 namespace/s
  2.80s: Benchmark on 4096 namespaces complete: 1464.3 namespace/s
  5.48s: Benchmark on 8192 namespaces complete: 1495.9 namespace/s
 11.19s: Benchmark on 16384 namespaces complete: 1464.7 namespace/s
 22.80s: Benchmark on 32768 namespaces complete: 1436.9 namespace/s
 45.33s: Benchmark on 65536 namespaces complete: 1445.8 namespace/s
 93.02s: Benchmark on 131072 namespaces complete: 1409.1 namespace/s
184.98s: Benchmark on 262144 namespaces complete: 1417.2 namespace/s
```

At the lower end, minor incremental slowdown (i assume some extra allocations and such), but at the higher end, wow! Note the relatively good scaling, with namespace rate staying relatively constant.

Using test script from #3630

`>for i in 256 512 1024 2048 4096 8192 16384 32768; do src/cmd/flux start ./test3633.sh -m unlimited -n $i; done | grep Script`

Before

```
Script runtime:          1.489s
Script runtime:          2.689s
Script runtime:          5.637s
Script runtime:         11.162s
Script runtime:         22.577s
Script runtime:         48.433s
Script runtime:        114.042s
Script runtime:        255.004s
```

After

```
Script runtime:          1.536s
Script runtime:          2.713s
Script runtime:          5.567s
Script runtime:         11.225s
Script runtime:         22.400s
Script runtime:         45.295s
Script runtime:        102.272s
Script runtime:        218.626s
```

Again at the higher ends, it's a nice chunk better.  

I originally did this experiment thinking to replace the setroot event's array with an object.  So that the extra allocations wouldn't occur and an array search wouldn't occur on setroot listeners.

However, I'd feel remiss to change the RPC to set an object over an array, where the object values are completely useless.  And we don't technically have a bottleneck in the array search on setroot listeners.  This one location is the only bottleneck i know of.

So I'm thinking this experiment may be a decent enough win for the time being.

Although using a `json_object()` as a cheapo hash is sort of, well silly.  Especially since we don't need the value of the keys we set.  Unfortunately, I think it's probably better than a `zhash`.    We don't seem to have a `set` data structure laying around in `flux-core`.  I looked at @grondo's `grudgeset` but that uses a `json_object` internally already.

Will peruse the interwebs to see if we can find a better one.  But wanted to post this for now if anyone can think of a `set` data structure out there we can use that would be good.