@garlick 2021-05-16
Yes!  Nice!

> Will peruse the interwebs to see if we can find a better one. But wanted to post this for now if anyone can think of a set data structure out there we can use that would be good.

Maybe this ccan module?

http://ccodearchive.net/info/objset.html

It's pretty easy to pull in more  libccan classes if they are useful and they have the right license.  This one's LGPL so should be fine.