@grondo 2021-05-16
If the intent is just to determine if a string is a member of a set, I wonder if a trie would be fastest. Since (I think, definitely not an expert) a trie will abort on the first non-matching character it should be pretty fast in this situation, and also very compact when many keys share a prefix (common for the KVS?). Though I guess this is closer to `strcmp()` so probably I don't know what I'm talking about.

ccan doesn't have a trie so that leads me to believe my suggestion is naive. (There is a [triemap](https://github.com/RedisLabs/triemap) as part of Redis, but the interface looks a bit funny)

ccan does have [strmap](https://ccodearchive.net/info/strmap.html) and [strset](https://ccodearchive.net/info/strset.html) which are interesting because they are implemented with [critbit trees](https://cr.yp.to/critbit.html), which sound interesting. (However, there's a warning that

> ccan/htable is faster and uses less memory, but doesn't provide ordered or prefix operations.

So if you don't need ordered operations that doesn't sound like the way to go. (I only mention it because it was interesting)

Sorry if my musings are a bit of a waste of time!