@garlick 2021-05-16
`strset` actually looks fairly appealing for this!

However, backing away a bit, I think this setroot keys array is only used by the KVS watch implementation.  Upon receipt of a setroot event, the `kvs-watch` module iterates over all watchers in that namespace, checking if they are watching a key in that array.  Neither the order nor the uniqueness appears to matter.  Am I reading that right @chu11?

So one trivial thing to try would be to just append a key to the array without checking uniqueness and see how that goes. I'm not clear under what circumstances duplicates occur, but if it's infrequent, maybe any growth of the array would not be a concern.

Another would be to convert the keys array to a json dict (say using json null as the value since it doesn't require a malloc when encoding/decoding).  This might simplify things all around.