@chu11 2021-05-16
> strset actually looks fairly appealing for this!

It seems like this could work.  When I first looked at it I thought that the user was supposed to supply memory for the string.  But looking at it again, it doesn't.  Perhaps I was looking at a different lib and got confused :P

> So one trivial thing to try would be to just append a key to the array without checking uniqueness and see how that goes. I'm not clear under what circumstances duplicates occur, but if it's infrequent, maybe any growth of the array would not be a concern.

Had considered this as well, although haven't experimented with it yet.  I figured it would be lots of duplicates, b/c there would be lots of appends to logs and stuff.  Although I didn't verify this.

> Another would be to convert the keys array to a json dict (say using json null as the value since it doesn't require a malloc when encoding/decoding). This might simplify things all around.

That was my original plan, but it felt a little wrong to do this.  I'd be modifying the RPC and populating it with effectively useless information.  But the fact you bring it up maybe means its not as offensive as I made it out to be :P  I'll try it out and see how it looks.