@garlick 2021-05-16
> I figured it would be lots of duplicates, b/c there would be lots of appends to logs and stuff. Although I didn't verify this.

Great point, I wasn't thinking about that particular case.  You are probably right.

> I'd be modifying the RPC and populating it with effectively useless information.

I think the protocol is OK to break at this point, and the extra `:null` per encoding seems pretty innocuous to me.  The use of a dict does communicate that the entries are unique and unordered, so this is maybe more self-documenting.