@grondo 2021-05-16
Also just throwing out there another Redis project I stumbled across:

 https://github.com/antirez/rax

Sounds like use of a trie/radix tree isn't really necessary here. However, for future reference, this kind of data structure _might_ be faster than a hash in certain circumstances. For example when you need to determine if a relatively large number of strings are in a small set, since then you avoid the need to call a hash function on every tested string, and the tested string is rejected as soon as the first character doesn't match a prefix of any string stored in the tree. (of course, maybe this would never be the case for any of our use cases)