@chu11 2021-05-16
> The use of a dict does communicate that the entries are unique and unordered, so this is maybe more self-documenting.

Ahh good point, hadn't thought of it that way.