@chu11 2021-05-17
Re-pushed, using a json dict everywhere including the RPC, rather than just for the "hash" lookup of duplicates.

Initial implementation w/ just a json "hash" to detect duplicates (copy of the results posted above):

```
  0.19s: Benchmark on 256 namespaces complete: 1379.0 namespace/s
  0.37s: Benchmark on 512 namespaces complete: 1401.2 namespace/s
  0.71s: Benchmark on 1024 namespaces complete: 1437.1 namespace/s
  1.34s: Benchmark on 2048 namespaces complete: 1526.3 namespace/s
  2.80s: Benchmark on 4096 namespaces complete: 1464.3 namespace/s
  5.48s: Benchmark on 8192 namespaces complete: 1495.9 namespace/s
 11.19s: Benchmark on 16384 namespaces complete: 1464.7 namespace/s
 22.80s: Benchmark on 32768 namespaces complete: 1436.9 namespace/s
 45.33s: Benchmark on 65536 namespaces complete: 1445.8 namespace/s
 93.02s: Benchmark on 131072 namespaces complete: 1409.1 namespace/s
184.98s: Benchmark on 262144 namespaces complete: 1417.2 namespace/s
```

```
Script runtime:          1.536s
Script runtime:          2.713s
Script runtime:          5.567s
Script runtime:         11.225s
Script runtime:         22.400s
Script runtime:         45.295s
Script runtime:        102.272s
Script runtime:        218.626s
```

New implementation where RPC uses json dict instead of array

```
  0.18s: Benchmark on 256 namespaces complete: 1441.7 namespace/s
  0.35s: Benchmark on 512 namespaces complete: 1453.2 namespace/s
  0.69s: Benchmark on 1024 namespaces complete: 1488.3 namespace/s
  1.38s: Benchmark on 2048 namespaces complete: 1480.0 namespace/s
  2.80s: Benchmark on 4096 namespaces complete: 1464.5 namespace/s
  5.58s: Benchmark on 8192 namespaces complete: 1467.6 namespace/s
 10.60s: Benchmark on 16384 namespaces complete: 1546.3 namespace/s
 22.41s: Benchmark on 32768 namespaces complete: 1462.1 namespace/s
 44.60s: Benchmark on 65536 namespaces complete: 1469.4 namespace/s
 90.33s: Benchmark on 131072 namespaces complete: 1451.1 namespace/s
179.93s: Benchmark on 262144 namespaces complete: 1456.9 namespace/s
```

```
Script runtime:          1.523s
Script runtime:          2.870s
Script runtime:          5.531s
Script runtime:         11.363s
Script runtime:         21.768s
Script runtime:         44.236s
Script runtime:         95.159s
Script runtime:        206.803s
```

gained a little bit more with this implementation.