@chu11 2021-05-17
coverage is a little low, lemme see if I can add a unit test to fix that.  i think I need a merge of 3 transactions instead of 2 to get that extra coverage in a block.