@chu11 2021-05-17
re-pushed with the minor issues found by @garlick above.  Also realized the low coverage was b/c one chunk of code is actually dead code.  So added a commit to remove that dead code.