@garlick 2021-05-17
Good find on the dead code.  Feel free to set MWP when you're ready.